django_backend/requirements.txt
asgiref==3.6.0
boto3==1.40.45
botocore==1.40.45
dj-database-url==3.0.1
Django==4.2
django-cors-headers==3.14.0
django-storages==1.14.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
Pillow==9.5.0
psycopg2-binary==2.9.10
PyJWT==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2023.3
s3transfer==0.14.0
six==1.17.0
sqlparse==0.4.3
urllib3==2.5.0